# internal/http/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.signinRequest:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  auth.signupRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.ErrorBody'
      meta: {}
      request_id:
        type: string
      success:
        type: boolean
    type: object
  httpx.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: MrChat backend API for auth, user settings, chat completions, billing,
    and admin configuration. Swagger UI is exposed at /swagger/index.html.
  title: MrChat API
  version: "0.1"
paths:
...
  /auth/refresh:
    post:
      description: Refresh access token by refresh cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Refresh access token
      tags:
      - Auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate by username or email and issue access/refresh tokens
      parameters:
      - description: Signin payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.signinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Sign in
      tags:
      - Auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account and issue access/refresh tokens
      parameters:
      - description: Signup payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Sign up
      tags:
      - Auth
  /billing/logs:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Billing log type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: List billing logs
      tags:
      - Billing
  /billing/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Get billing summary
      tags:
      - Billing
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter access token as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"